[refactor] page/ コンポーネントの store 由来 props を子側購読へ (WP-H6 PR4)
- PrimaryWorkspace: profileAuthorLabel を props から外し、購読済みの
  syncStatus / localProfile から子側で算出
- NotificationsWorkspace: notificationItems の組み立て(useMemo)を
  Page から移設し、knownAuthorsByPubkey / mediaObjectUrls を子側購読に
- DetailPaneStack / Overlays: activeTopic を子側 selector 購読に
- SettingsDrawer: drawerId を SHELL_SETTINGS_ID の直接 import に
  (AuxiliaryPanels の SHELL_CONTEXT_ID と同型)
- selectShellPageSlice から localProfile / mediaObjectUrls /
  knownAuthorsByPubkey を削除。メディア読込・作者情報の更新で
  Page(木の根)が再レンダーされなくなる
- 残る props はフック単一インスタンスの出力(viewModels / handler 群)、
  Page ローカル useState / ref、環境(t / locale / api)、合成用 ReactNode
  のみ(本質依存)
- DesktopShellPage 1153 → 1090 行(oversized baseline を ratchet down)
